some stuff

+++ src/Data/ImageData.java
@@ -3,12 +3,12 @@ package Data;
 public class ImageData {
 	
 	// Array of Gray pixel Values 0 - 255
-	private int grayValues[]; 
+	private float grayValues[]; 
 	// Classification of an Image 0-9
 	private int label;
 	
 	// Getters and Setters
-	public int[] getGrayValues() {
+	public float[] getGrayValues() {
 		return grayValues;
 	}
 	public int getLabel() {
@@ -16,7 +16,7 @@ public class ImageData {
 	}
 	
 	
-	public ImageData(int[] grayValues,int label ) {
+	public ImageData(float[] grayValues,int label ) {
 		this.label = label;
 		this.grayValues = grayValues;
 	}
@@ -26,7 +26,7 @@ public class ImageData {
 	{
 		StringBuilder sb = new StringBuilder();
 		sb.append(this.label+ "\n\n");
-		for(int grayValue :this.grayValues)
+		for(float grayValue :this.grayValues)
 			sb.append(grayValue+ " ");
 		return sb.toString();
 	}
